Break out component props into separate files
