+++ packages/combobox/src/index.ts
@@ -1,4 +1,4 @@
 export { Combobox } from './Combobox';
 export { ComboboxGroup } from './ComboboxGroup';
 export { ComboboxOption } from './ComboboxOption';
-export { type ComboboxProps } from './types/Combobox.types';
+export { type ComboboxProps } from './types';
