+++ packages/combobox/src/types/index.ts
@@ -1,0 +1,23 @@
+export {
+  ComboboxElement,
+  ComboboxSize,
+  getNullSelection,
+  onChangeType,
+  Overflow,
+  SearchState,
+  SelectValueType,
+  State,
+} from './Combobox.types';
+export { ComboboxChipProps } from './ComboboxChipProps.types';
+export { ComboboxGroupProps } from './ComboboxGroupProps.types';
+export {
+  ComboboxOptionProps,
+  InternalComboboxOptionProps,
+  OptionObject,
+} from './ComboboxOptionProps.types';
+export {
+  BaseComboboxProps,
+  ComboboxMultiselectProps,
+  ComboboxProps,
+} from './ComboboxProps.types';
+export { TruncationLocation } from '@leafygreen-ui/chip';
